backend/requirements.txt
# ─────────────────────────────────────────────────────────────
# OrderBookExpert — Backend Requirements
# FastAPI + TimescaleDB + JWT + Celery
# ─────────────────────────────────────────────────────────────

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database (async SQLAlchemy + TimescaleDB/PostgreSQL)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

# Pydantic Settings (env var management)
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# Task Queue
celery==5.4.0
redis==5.1.1

# WebSocket support (included with FastAPI/uvicorn)
websockets==13.1

# HTTP Client (for exchange API calls)
httpx==0.27.2

# Exchange integration
ccxt==4.4.22

# Date & Time utilities
python-dateutil==2.9.0

# Machine Learning & Data Science
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
